+++ core/routes/index.js
@@ -4,6 +4,7 @@ import { setupNotesRoutes } from "./notes.js"
 import { setupTagCloudRoute } from "./tag-cloud.js"
 import { setupVideoLibraryRoute } from "./videos.js"
 import { setupSearchRoute } from "./search.js"
+import { setupStageRoutes } from "./stage.js"
 import { setupTaxonomyRoutes } from "./taxonomy.js"
 import { setupCustomRoutes } from "./custom.js"
 import { setupSeoRoutes } from "./seo.js"
@@ -16,6 +17,7 @@ export function setupFrontendRoutes(app, systems) {
     setupTagCloudRoute(app, systems)
     setupVideoLibraryRoute(app, systems)
     setupSearchRoute(app, systems)
+    setupStageRoutes(app, systems)
     setupTaxonomyRoutes(app, systems)
     setupCustomRoutes(app, systems)
     setupSeoRoutes(app, systems)